backend/requirements.txt
pdfplumber
docx2txt
pymupdf
python-docx
